requirements.txt
PyQt6>=6.4.0
mss>=9.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
